requirements.txt
oauth2client >= 4.1.3
google-api-python-client >= 1.12.5
google-auth-httplib2 >= 0.0.4
google-auth-oauthlib >= 0.4.2
pandas
Pyarrow